Use 64-bit integer for chunk ID
* Using a 32-bit integer, signed, gives only 2 billion chunks before it wraps.
  With Vulkan/D3D12 creating new chunks even while in the background for
  recording commands this is feasible to hit.

+++ renderdoc/driver/d3d11/d3d11_device.cpp
@@ -1975,7 +1975,7 @@ bool WrappedID3D11Device::EndFrameCapture(void *dev, void *wnd)
 
         RDCDEBUG("Accumulating context resource list");
 
-        std::map<int32_t, Chunk *> recordlist;
+        std::map<int64_t, Chunk *> recordlist;
         record->Insert(recordlist);
 
         RDCDEBUG("Flushing %u records to file serialiser", (uint32_t)recordlist.size());
