+++ renderdoc/driver/d3d12/d3d12_device.cpp
@@ -2104,7 +2104,7 @@ bool WrappedID3D12Device::EndFrameCapture(void *dev, void *wnd)
     // in capframe (the transition is thread-protected) so nothing will be
     // pushed to the vector
 
-    std::map<int32_t, Chunk *> recordlist;
+    std::map<int64_t, Chunk *> recordlist;
 
     for(auto it = queues.begin(); it != queues.end(); ++it)
     {
